// client/src/models/weather.ts
export interface Weather {
  location: WeatherLocation;
  current: CurrentWeather;
  forecast: {
    forecastday: Array<WeatherForecast>;
  };
}
// ...
export interface WeatherLocation {
  name: "London";
  region: "City of London, Greater London";
  country: "United Kingdom";
  lat: 51.52;
  lon: -0.11;
  tz_id: "Europe/London";
  localtime_epoch: 1655142120;
  localtime: "2022-06-13 18:42";
}

export interface CurrentWeather {
  last_updated_epoch: 1655141400;
  last_updated: "2022-06-13 18:30";
  temp_c: 20;
  temp_f: 68;
  is_day: 1;
  condition: {
    text: "Sunny";
    icon: "//cdn.weatherapi.com/weather/64x64/day/113.png";
    code: 1000;
  };
  wind_mph: 2.2;
  wind_kph: 3.6;
  wind_degree: 336;
  wind_dir: "NNW";
  pressure_mb: 1024;
  pressure_in: 30.24;
  precip_mm: 0;
  precip_in: 0;
  humidity: 43;
  cloud: 0;
  feelslike_c: 20;
  feelslike_f: 68;
  vis_km: 10;
  vis_miles: 6;
  uv: 6;
  gust_mph: 7.4;
  gust_kph: 11.9;
}